requirements.txt
audioop-lts==0.2.1
AwesomeTkinter==2021.11.8
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
PyAudio==0.2.14
python-bidi==0.6.6
pyzbar==0.1.9
qrcode==8.0
screeninfo==0.8.1
SpeechRecognition==3.14.1
standard-aifc==3.13.0
standard-chunk==3.13.0
typing_extensions==4.12.2
